+++ base/export/auth-helpers.d.ts
@@ -1,1 +1,0 @@
-export { firebaseConfig, application, db, storage, auth, user } from '../firebase/authentication';