+++ base/export/auth-helpers.js
@@ -1,2 +1,0 @@
-export { firebaseConfig, application, db, storage, auth, user } from '../firebase/authentication';
-//# sourceMappingURL=auth-helpers.js.map