+++ base/firebase/interface.d.ts
@@ -1,4 +1,0 @@
-export declare const logOut: () => Promise<void>;
-export declare const logAccount: () => Promise<void>;
-export declare const logAccountDelete: () => Promise<void>;
-export declare const deleteDocument: (collect: string, item: string) => Promise<void>;