+++ base/webcomponents/drawer/css/toggle.d.ts
@@ -1,2 +1,0 @@
-import { CSSResult } from "lit";
-export declare const toggleStyle: CSSResult;