+++ base/webcomponents/icon/css/icon.d.ts
@@ -1,1 +1,0 @@
-export declare const iconStyles: import("lit").CSSResult;