+++ base/webcomponents/settings/css/navigation.d.ts
@@ -1,1 +1,0 @@
-export declare const navigation: import("lit").CSSResult;