+++ export/pwa-auth.ts
@@ -1,4 +1,5 @@
-export { UserDrawer }           from '../base/webcomponents/drawer/drawer';
-export { UserSettings }         from '../base/webcomponents/settings/settings';
-export { UserIcon }             from '../base/webcomponents/icon/icon';
-export { firebaseConfig, application, db, storage, auth, user } from '../base/firebase/authentication';
+export { UserDrawer }                   from '../bundle/base/webcomponents/drawer/drawer';
+export { UserSettings }                 from '../bundle/base/webcomponents/settings/settings';
+export { UserIcon }                     from '../bundle/base/webcomponents/icon/icon';
+export { store }                        from '../bundle/base/store';
+export { db, storage, auth, user }      from '../bundle/base/firebase/authentication';