+++ CHANGELOG.md
@@ -6,9 +6,15 @@ All notable changes to pwa-auth.js will be documented in this file.
 - Social Login Flow
 - animated login and data
 
+# [3.2.4] - 2025-10-08
+
+## Added
+- Export Redux State to Public
+- webcomponent: Login Widget
+
 # [3.2.3] - 2025-10-08
 
-## Changed
+## Added
 - Rollup Project Bundle
 
 # [3.2.1] - 2025-10-07